+++ stories/atoms/SelectInput.stories.tsx
@@ -1,6 +1,12 @@
 import type { Meta, StoryObj } from '@storybook/react'
 import { SelectInput } from '@hce/design-system'
 
+const EMPRESAS = [
+  { value: "CENTRAL", label: "Sede Central" },
+  { value: "NORTE",   label: "Sede Norte"   },
+  { value: "SUR",     label: "Sede Sur"     },
+]
+
 const meta: Meta<typeof SelectInput> = {
   title: 'Atoms/SelectInput',
   component: SelectInput,
@@ -18,5 +24,6 @@ export const Default: Story = {
     fullWidth: false,
     size: "small",
     required: false,
+    options: EMPRESAS
   },
 }